+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
 import { store } from './store/store'
-import { createSpriteService } from './sprites/service'
+import { createSpriteServiceV2 } from '@/sprites/service'
 import App from './app/App'
 import './index.css'
 
@@ -10,7 +10,7 @@ import './index.css'
 async function initializeApp(): Promise<void> {
   try {
     // Load sprites before initializing Redux
-    const spriteService = await createSpriteService()
+    const spriteService = await createSpriteServiceV2()
 
     // Now render the app with sprites loaded
     createRoot(document.getElementById('root')!).render(
